# feat(gatsby): upgrade graphql-compose to latest for speed improvements