--- a/packages/gatsby/package.json
+++ b/packages/gatsby/package.json
@@ -91,7 +91,7 @@
     "glob": "^7.2.0",
     "got": "^11.8.2",
     "graphql": "^15.7.2",
-    "graphql-compose": "~7.25.1",
+    "graphql-compose": "^9.0.6",
     "graphql-playground-middleware-express": "^1.7.22",
     "hasha": "^5.2.2",
     "http-proxy": "^1.18.1",