<!--
  Have any questions? Check out the contributing docs at https://gatsby.dev/contribute, or
  ask in this Pull Request and a Gatsby maintainer will be happy to help :)

  For any major changes, please first open a bug report (if it's a bug) or a feature request.
-->

## Description

<!-- Write a brief description of the changes introduced by this PR -->
Large nested queries take a long time. Most time is spent inside graphql-compose as it fetches type information for every field for each query.
A lot of data is being passed around and it causes lots of GC.

The latest grahpql-compose version does memoisation of getTypes increases drops performance of these queries dramatically.
https://github.com/graphql-compose/graphql-compose/releases/tag/v9.0.0

You can try the canary with `gatsby@graphql-compose`

![image](https://user-images.githubusercontent.com/1120926/149678034-e9748ae2-9d44-4cc1-b35a-63f173f0d928.png)

Some numbers to share:

<details>
<summary>customer 1</summary>

before:
```
success building schema - 256.133s
success createPages - 993.229s
success createPagesStatefully - 0.094s
```

after:
```
success building schema - 181.002s
success createPages - 156.464s
success createPagesStatefully - 0.144s
```
</details>

<details>
<summary>customer 2</summary>

before:
```
success building schema - 358s
```

after:
```
success building schema - 90s
```
</details>

## Related Issues

[sc-42224]